mtf2json: refactored MTF file parsing
Parsing is now done by a generator function returning a (key, value,
section) tuple. In case of a critical slot line containing ':size:' or
':SIZE:', the value is set to the preceding string (the size part is
discarded). This fixes the issue described in #15 and creates the
desired result in the 'critical_slots' section. However, it does not add
the cargo, liquid_storage and communication equipment to the list of
weapons, as MegaMek does when creating record sheets.

Also added the 'verbose' flag.