ignored keys: added 'nocrit' to ignored keys
According to the devs, that key was intended to describe equipment that
is available but does not occupy a critical slot. In some MTF files it
is used for the internal (standard) structure. However, it's not
actually used wihin MegaMek or for record sheet creation. If you open
e.g. Powerman_SC_XVI_HaulerMech.mtf in MegaMekLab and save it, the
`nocrit` entries disappear.